remove bugs assocaited with keeping track of pulse height in Redis and overflow bug in noise RMS calculation. Restore thresholds to be a float to be on the safe side

+++ sbndcode/DaqAnalysis/Analysis.h
@@ -111,7 +111,7 @@ public:
     bool redis;
     int n_headers;
 
-    int16_t threshold;
+    float threshold;
     float threshold_sigma;
     
     unsigned baseline_calc;
